chore: move sn_api_to_cairo_vm_program into snapi
