+++ crates/starknet_api/Cargo.toml
@@ -11,7 +11,9 @@ testing = []
 
 [dependencies]
 bitvec = "1.0.1"
+cairo-lang-casm.workspace = true
 cairo-lang-starknet-classes = "2.7.0-dev.0"
+cairo-vm.workspace = true
 derive_more = "0.99.17"
 hex = "0.4.3"
 indexmap = { version = "2.1.0", features = ["serde"] }
